This commit includes a comprehensive refactoring and improvement of the entire codebase.
Key changes include:

Backend:
- Fixed a critical bug in the server error handling middleware in `server/index.ts` that caused the server to crash on errors.
- Refactored the backend routing from a single monolithic file into a modular, domain-driven structure under `server/routes/`.
- Made the Express application instance in `server/index.ts` exportable to facilitate testing.

Frontend:
- Created a reusable `AppLayout` component to handle the layout for protected pages.
- Refactored `client/src/App.tsx` to use the new `AppLayout` component, significantly reducing code duplication and improving maintainability.

Testing:
- Introduced `vitest` as the testing framework for both frontend and backend.
- Added and configured necessary testing libraries, including `@testing-library/react`, `jsdom`, and `supertest`.
- Created a `vitest` workspace to manage the separate testing environments for the client and server.
- Wrote an initial backend integration test for the `/health` endpoint.
- Wrote an initial frontend component test for the `Landing` page.

Note: The final step of running the tests failed due to a `vitest` workspace configuration issue. The test runner was unable to resolve the paths to the configuration files correctly and indicated that workspace files are deprecated. Further work is needed to migrate this configuration to the root `vite.config.ts` to get the tests running.
