+++ client/src/components/layout/AppLayout.tsx
@@ -1,0 +1,19 @@
+import Sidebar from "@/components/layout/sidebar";
+import React from "react";
+
+interface AppLayoutProps {
+  children: React.ReactNode;
+}
+
+const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
+  return (
+    <div className="flex h-screen overflow-hidden">
+      <Sidebar />
+      <main className="flex-1 flex flex-col overflow-hidden">
+        {children}
+      </main>
+    </div>
+  );
+};
+
+export default AppLayout;
