+++ vitest.workspace.ts
@@ -1,0 +1,6 @@
+import { defineWorkspace } from 'vitest/config';
+
+export default defineWorkspace([
+  'client/vitest.config.ts',
+  'server/vitest.config.ts',
+]);
